adopet/serializers.py
from rest_framework import serializers
from adopet.models import Tutor

class TutorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tutor
        fields = ['id', 'name', 'email', 'password']
        