requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.2
Markdown==3.7
Pygments==2.18.0
sqlparse==0.5.1
tzdata==2024.2
